requirements.txt
asgi-redis==0.14.1
asgiref==0.14.0
autobahn==0.16.0
backports.ssl-match-hostname==3.5.0.1
channels==0.17.2
daphne==0.15.0
dask==0.11.1
decorator==4.0.10
Django==1.10.1
msgpack-python==0.4.8
pika==0.10.0
pyparsing==2.1.10
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
redis==2.10.5
scipy==0.17.1
six==1.10.0
Twisted==16.4.1
txaio==2.5.1
ujson==1.35
websocket-client==0.37.0
zope.interface==4.3.2
psycopg2==2.7.1
requests==2.14.2
